Port to BePinEx mod framework.
* AnimateObjectPatches: directly access m_optionScalar instead of using
FieldRef.
* CardCountManager: make CardCounts an injectable class, add IntPtr
constructor to CardCountManager.
* CardTabBehaviour: add IntPtr constructor to CardTabBehaviour, adapt
Add/RemoveListener calls to IL2CPP, use accessors for CardCounts
members.
* GameEventHandler: cast IntPtr to Int64 instead of Int32, since TS is
now 64-bit.
* GameLogPatches: use BePinEx logger instead of .NET ILogger, replace
Select() with a for loop because IL2CPP collection classes don't work
with Linq.
* RegionControlBar: add IntPtr constructor to RegionControlBar, use
ScorePanel.AnimateBar instead of a locally-defined coroutine because
managed coroutines don't work with IL2CPP.
* TwilightLibWrapper: change GetGameFinalScoreState to construct the
GameFinalRegionScoreStates one at a time because GFSS is no longer a
"formatted" class, make Players an injectable class (needed for the
CardCountManager changes).
* TwilightStrugglePatches: replace typeof with Il2CppType.Of when
constructing RegionControlBars, add interop method for Sprite.Create
because BePinEx's automated generation doesn't handle it.
* UI_OnlineCreateGamePatches: access m_delayCoroutine through the
__instance parameter instead of a separate private member parameter
because it's no longer declared as private.
* UI_SettingsMenuPatches: create an injectable SettingsCallbacks class
and use it for the HandleSceneLoaded callback.
* InitEspionage: rename to TsEspionage, register injectable classes.
* TSEspionage.csproj: update to work with BePinEx, keep interop DLLs in
the Twilight Struggle install directory, add Test and Release targets.
* Update installation instructions and screenshot.

+++ TSEspionage/GameEventHandler.cs
@@ -28,7 +28,7 @@ namespace TSEspionage
 
         public void HandleEvent(ref IntPtr eventBuffer)
         {
-            var eventType = (EventType)Marshal.ReadIntPtr(eventBuffer).ToInt32();
+            var eventType = (EventType)Marshal.ReadIntPtr(eventBuffer).ToInt64();
             var eventPointer = eventBuffer + Marshal.SizeOf(typeof(int));
 
             switch (eventType)
